segundo commit

+++ Models/Productos.cs
@@ -4,7 +4,7 @@ namespace tl2_tp8_2025_Clari002.Models
     {
         public int IdProducto { get; set; }
         public string? Descripcion { get; set; }
-        public int Precio { get; set; }
+        public double Precio { get; set; }
     }
     
 }
